+++ src/vcs/grep.cpp
@@ -6,13 +6,15 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include <boost/algorithm/string.hpp>
+#include <wex/core/config.h>
 #include <wex/core/log.h>
-#include <wex/stc/entry-dialog.h>
+#include <wex/factory/process-data.h>
 #include <wex/stc/stc.h>
 #include <wex/ui/frame.h>
+#include <wex/ui/frd.h>
+#include <wex/ui/item-build.h>
+#include <wex/ui/item-dialog.h>
 #include <wx/app.h>
-
-#include "util.h"
 
 namespace wex
 {
@@ -21,28 +23,27 @@ bool execute_grep(const std::string& bin, const path& tl)
   auto* frame = dynamic_cast<wex::frame*>(wxTheApp->GetTopWindow());
   auto* stc   = frame->get_stc();
   auto  text(stc->get_selected_text());
-  static stc_entry_dialog* dlg = nullptr;
+  static item_dialog* dlg = nullptr;
 
   if (text.empty())
   {
     if (dlg == nullptr)
     {
-      dlg = new stc_entry_dialog(
-        std::string(),
-        _("Text") + ":",
-        wex::data::window().title("git grep"),
-        wex::data::stc().flags(
-          wex::data::stc::window_t().set(wex::data::stc::WIN_SINGLE_LINE)));
-    }
+      const item::choices_bool_t checks{
+        find_replace_data::get()->text_match_case(),
+        find_replace_data::get()->text_match_word()};
 
-    dlg->get_stc()->SetFocus();
+      dlg = new item_dialog(
+        {add_find_text(), checks},
+        wex::data::window().title("git grep"));
+    }
 
     if (dlg->ShowModal() == wxID_CANCEL)
     {
       return false;
     }
 
-    text = dlg->get_stc()->get_text();
+    text = find_replace_data::get()->get_find_string();
 
     if (text.empty())
     {
@@ -51,7 +52,8 @@ bool execute_grep(const std::string& bin, const path& tl)
   }
   else if (dlg != nullptr)
   {
-    dlg->get_stc()->set_text(text);
+    stc->get_find_string();
+    dlg->reload();
   }
 
   if (const std::string &
@@ -62,8 +64,17 @@ bool execute_grep(const std::string& bin, const path& tl)
   }
   else
   {
+    const std::string flag(
+      config(find_replace_data::get()->text_match_case()).get(true) ? "" :
+                                                                      " -i ");
+
+    const std::string finds(
+      config(find_replace_data::get()->text_match_word()).get(true) ?
+        "\\b" + find + "\\b" :
+        find);
+
     frame->process_async_system(
-      process_data(bin + " grep -n " + find).start_dir(tl.string()));
+      process_data(bin + " grep -n " + flag + finds).start_dir(tl.string()));
   }
 
   return true;